Made some minor fixes after the first mentor's review
